+++ opendut-lea/src/components/buttons/mod.rs
@@ -2,3 +2,4 @@ pub mod confirmation_button;
 pub mod doorhanger_button;
 pub mod icon_button;
 pub mod button;
+pub mod collapse_button;